Added a simulation mode that shows the user what commands would be run instead of actually performing the operation

+++ RDiskImager/src/RDiskImager.rb
@@ -13,8 +13,21 @@
 
 class RDiskImager
   def initialize()
-     testCompatibility()
-     getChoice()
+    testArgs()
+    testCompatibility()
+    getChoice()
+  end
+  
+  # See what command line arguments the user has set
+  def testArgs()
+    ARGV.each do |arg|
+      if (arg == "--simulate" or arg == "-s")
+        @simulate = true
+        puts "Running in simulation mode!"
+      else
+        @simulate = false
+      end
+    end
   end
   
   # make sure programs and files RDiskImager needs are present
@@ -46,7 +59,7 @@ class RDiskImager
   #   to an image or an image to a device
   def getChoice()
     puts "Device to Image [1]\nImage to Device [2]\n"
-    choice = gets().chomp
+    choice = $stdin.gets.chomp
     if choice == "1"
       deviceToImageDialog()
     end
@@ -58,7 +71,7 @@ class RDiskImager
   # Prompts user for the device and image names
   def deviceToImageDialog()
     puts "Device name? (sda, hdb, sdc, etc...)"
-    device_name = gets.chomp
+    device_name = $stdin.gets.chomp
         
     # read the device's size from /proc/partitions
     devtable = `cat /proc/partitions | grep "#{device_name}"`
@@ -68,9 +81,9 @@ class RDiskImager
     size = devtable.split("\n")[0].split(" ")[2]
         
     puts "Image name?"
-    image_name = gets.chomp
+    image_name = $stdin.gets.chomp
     puts "Are you sure you want to save an image of #{device_name} to #{image_name}? y/[n]"
-    if gets.chomp == "y"
+    if $stdin.gets.chomp == "y"
       deviceToImage(device_name, size, image_name)
     else
       puts "Cancelling operation!"
@@ -80,13 +93,13 @@ class RDiskImager
   # Prompts user for image and device names
   def imageToDeviceDialog()
     puts "Image name?"
-    image_name = gets.chomp
+    image_name = $stdin.gets.chomp
     size = File.size(image_name)/1024
     puts size
     puts "Device name? (sda, hdb, sdc, etc...)"
-    device_name = gets.chomp
+    device_name = $stdin.gets.chomp
     puts "Are you sure you want to write #{image_name} to #{device_name}? y/[n]"
-    if gets.chomp == "y"
+    if $stdin.gets.chomp == "y"
       imageToDevice(image_name, size, device_name)
     else
       puts "Cancelling operation!"
@@ -95,12 +108,22 @@ class RDiskImager
   
   # Write a device to an image
   def deviceToImage(device_name, size, image_name)
-    system "dd if=\"/dev/#{device_name}\"| pv -s #{size}k | dd of=\"#{image_name}\""
+    if @simulate
+      puts "This would run:"
+      puts "dd if=\"/dev/#{device_name}\"| pv -s #{size}k | dd of=\"#{image_name}\""
+    else
+      system "dd if=\"/dev/#{device_name}\"| pv -s #{size}k | dd of=\"#{image_name}\""
+    end
   end
   
   # Write an image to a device
   def imageToDevice(image_name, size, device_name)
-    system "dd if=\"#{image_name}\"| pv -s #{size}k | dd of=\"/dev/#{device_name}\""
+    if @simulate
+      puts "This would run:"
+      puts "dd if=\"#{image_name}\"| pv -s #{size}k | dd of=\"/dev/#{device_name}\""
+    else
+      system "dd if=\"#{image_name}\"| pv -s #{size}k | dd of=\"/dev/#{device_name}\""
+    end
   end
   
 end
